refactor!: separate "validate" into its own directory
