+++ raillabel/_understand_ai_t4_format/metadata.py
@@ -81,7 +81,9 @@ class Metadata:
         }
 
     def _get_subschema_version(self) -> str:
-        RAILLABEL_SCHEMA_PATH = Path(__file__).parent.parent / "schemas" / "raillabel_schema.json"
+        RAILLABEL_SCHEMA_PATH = (
+            Path(__file__).parent.parent / "validate" / "schemas" / "raillabel_schema.json"
+        )
 
         with RAILLABEL_SCHEMA_PATH.open() as schema_file:
             subschema_version = json.load(schema_file)["version"]
