+++ raillabel/format_loaders/loader_raillabel.py
@@ -24,7 +24,9 @@ class LoaderRailLabel(LoaderABC):
     scene: format.Scene
     warnings: t.List[str]
 
-    SCHEMA_PATH: Path = Path(__file__).parent.parent / "schemas" / "raillabel_schema.json"
+    SCHEMA_PATH: Path = (
+        Path(__file__).parent.parent / "validate" / "schemas" / "raillabel_schema.json"
+    )
 
     @property
     def subschema_version(self):
