+++ tests/conftest.py
@@ -90,7 +90,7 @@ import raillabel
 
 json_data_directories = [
     Path(__file__).parent / "test_raillabel" / "__test_assets__",
-    Path(__file__).parent.parent / "raillabel" / "schemas"
+    Path(__file__).parent.parent / "raillabel" / "validate" / "schemas"
 ]
 
 @pytest.fixture(scope="session", autouse=True)
